fix(ci-cd-code-review): use z.number() instead of z.number().int()
In Zod v4, .int() maps to safeint format which injects min/max of
±9007199254740991. The AI SDK passes these as minimum/maximum on the
integer type, which Anthropic's structured output rejects. Plain
z.number() emits no constraints; the prompt already enforces 1–10.

+++ packages/code-reviewer/src/schema.ts
@@ -9,7 +9,7 @@ export const criterionSchema = z.object({
     "documentation",
     "security_safety",
   ]),
-  score: z.number().int(),
+  score: z.number(),
   rationale: z.string(),
   issues: z.array(z.string()).optional(),
 });
@@ -17,7 +17,7 @@ export const criterionSchema = z.object({
 export const reviewSchema = z.object({
   summary: z.string(),
   criteria: z.array(criterionSchema),
-  overall_score: z.number().int(),
+  overall_score: z.number(),
 });
 
 export type ReviewOutput = z.infer<typeof reviewSchema>;
